+++ src/test/java/com/github/freenamu/parser/grammar/singleline/QuoteGrammarTest.java
@@ -1,9 +1,9 @@
 package com.github.freenamu.parser.grammar.singleline;
 
 import com.github.freenamu.parser.node.Node;
-import com.github.freenamu.parser.node.single.Bold;
-import com.github.freenamu.parser.node.single.Italic;
-import com.github.freenamu.parser.node.single.Text;
+import com.github.freenamu.parser.node.singleline.Bold;
+import com.github.freenamu.parser.node.singleline.Italic;
+import com.github.freenamu.parser.node.singleline.Text;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
